Activities: Add stage start_date to API response
The stage start_date wasn't included in the api response for
language activities. This left the stage start date field blank
on the list of activities on the initial load. However, if you
visited the activity details, it would load the start date into
the component and it would display on the language page after that.

This change adds the start date of the stage to the API call so
that dates are displayed immediately.