WIP: Tagging Events
This commit adds the concept of adding a tag to categorize events, events
concerning the same topic or same location can have the same tag.

This is not (yet, maybe) a general purpose tag that can be applied to any Dulu
model.

This may also only works on events per language not across language. The
interface to view other tagged events only will work per language with the
current implementation (and really only one language right now).

This is a suggestion from Drew.