Fix "Translation Missing" Fallback Errors
The i18n initializer would try to do a "fallback" to a missing
translation. This generally worked pretty well in, for instance,
a book of the bible where the english translation of Ruth is Ruth.

However, the bug is that the initializer tries to check if you're
asking for the english translation, then initiates the fallback.
But, it actually looks at the user's selected locale. This causes
issues with looking up english translations when the user's locale
is set to french, because then it has no fallback.

This changes the commit to look not at the user's locale, but the
requested locale when seeing if a fallback is appropriate (i.e.
English)

TODO, add cypress test which would allow testing, `rails test` skips
the ininitlizers so it cannot be tested in this way.