Participants Bug
Steps to Reproduce:
1. Visit language A with Participants who are attached to the
language only, not the langauage's cluster.
2. Using the navigation menu at the top, visit language B
which is itself not attached to a cluster.
3. See that language A's participants are displayed in
language B's people section on the main page.
4. Refresh the page and see that language A's people now
dissapear.

Summary
The participants filter matched on participant.cluster who has
no cluster and a language which has no cluster because they
are both null.

So the fix is that if a participant has no cluster, skip the
attempt to compare this participant's cluster with the
language's cluster and thus don't end up comparing null to null.

Testing
I tried to test this every which way but was unsuccessful. This
is why there is no failing test to demonstrate the behavior. When
I understand react testing better perhaps I will come back to this
(cypress didn't even work since it forces a refresh on navigating
to a new URL).