Update Email Sending Service
This is now using an external/cloud SMTP service. Previously this
was using a gmail account. Updated the config keys to reflect
this change, thus gmail_username and gmail_password are no longer
being used and now are generically smtp_username and smtp_password.

The new configs have already been placed on the production server
where the secrets.yml is kept, so it should begin working once the
change has been rolled to the server and the services restarted.